Initialize chat messages on the server
